[Chore/#175] 네트워크 Model 정리
