+++ Staccato-iOS/Staccato-iOS/Service/Login/DTO/LoginResponse.swift
