+++ Staccato-iOS/Staccato-iOS/Service/Protocol/HeaderType.swift
