--- Staccato-iOS/Staccato-iOS/Util/Model/Login/LoginRequest.swift
@@ -1,12 +1,0 @@
-//
-//  LoginRequest.swift
-//  Staccato-iOS
-//
-//  Created by 강재혁 on 1/9/25.
-//
-
-import Foundation
-
-struct LoginRequest: Codable {
-    let nickname: String            // 닉네임
-}
