--- Staccato-iOS/Staccato-iOS/Util/Model/Member/RecoveryCodeResponse.swift
@@ -1,12 +1,0 @@
-//
-//  RecoveryCodeResponse.swift
-//  Staccato-iOS
-//
-//  Created by 강재혁 on 1/9/25.
-//
-
-import Foundation
-
-struct RecoveryCodeResponse: Codable {
-    let token: String // 복구 코드 토큰
-}
