--- Staccato-iOS/Staccato-iOS/Util/Model/MyPage/AccountInformationResponse.swift
@@ -1,14 +1,0 @@
-//
-//  AccountInformationResponse.swift
-//  Staccato-iOS
-//
-//  Created by 강재혁 on 1/9/25.
-//
-
-import Foundation
-
-struct AccountInformationResponse: Codable {
-    var nickname: String             // 사용자 닉네임
-    var profileImageUrl: String?     // 프로필 이미지 URL (선택적)
-    var code: String                 // 사용자 코드
-}
