--- Staccato-iOS/Staccato-iOS/Util/Model/TimeLine/TimeLineMember.swift
@@ -1,13 +1,0 @@
-//
-//  TimeLineMember.swift
-//  Staccato-iOS
-//
-//  Created by 강재혁 on 1/9/25.
-//
-
-import Foundation
-
-struct TimelineMember: Codable {
-    let memberId: Int        // 구성원의 고유 ID
-    let memberImageUrl: String // 구성원의 이미지 URL
-}
