--- Staccato-iOS/Staccato-iOS/Util/Model/TimeLine/TimeLineResponse.swift
@@ -1,12 +1,0 @@
-//
-//  TimeLineResponse.swift
-//  Staccato-iOS
-//
-//  Created by 강재혁 on 1/9/25.
-//
-
-import Foundation
-
-struct TimelineResponse: Codable {
-    let memories: [TimelineMemory] // 메모리 목록
-}
